# services/api-gateway/src/main/resources/static/openapi/flowpay-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowPay API",
    "version": "v1",
    "description": "FlowPay is a digital wallet platform. This is the **public API**, reached through the gateway at\n`http://localhost:8080`.\n\n## Quick start\n1. **Register** - `POST /api/v1/users`\n2. **Log in** - `POST /api/v1/auth/login` -> copy the `accessToken`\n3. Click **Authorize** (top right) and paste the token. Protected endpoints (padlock icon) now work.\n4. **View / update your profile** - `GET` / `PATCH /api/v1/users/{id}`\n\n## How security works\n* Only registration, login, refresh, logout, password reset and the health check are public.\n* Everything else needs `Authorization: Bearer <accessToken>`. Tokens are JWTs signed with RS256 and last 15 minutes;\n  renew them with `POST /api/v1/auth/refresh` (refresh tokens are single use).\n* You only ever see and change **your own** data. Another customer's id returns `404`.\n* Several endpoints are rate limited per IP address and return `429` with a `Retry-After` header when exceeded.\n* Every error has the same shape: `{timestamp, status, error, message, path}`. Branch on `error`, not `message`.\n\n## Emails in development\nPassword reset and two-step verification send real emails to a local inbox (Mailpit): http://localhost:8025.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local gateway (docker compose)"
    }
  ],
...
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "2. Authentication"
        ],
        "summary": "Log in with email and password",
        "description": "Signs a customer in. **Public.** Limited to 10 attempts per minute per IP address (shared with `login/verify`).\n\n### How it works\n* Correct password, two-step verification **off** -> `200` with `accessToken` (JWT, 15 min) and `refreshToken` (7 days).\n* Correct password, two-step verification **on** -> `200` with `{\"mfaRequired\": true, \"challengeId\": ...}` and **no tokens**. A 6-digit code is emailed; finish with `POST /api/v1/auth/login/verify`.\n* Anything else -> the **same** `401 INVALID_CREDENTIALS` message for a wrong password, an unknown email and a locked account, so the API never reveals whether an account exists.\n* **5 wrong passwords lock the account for 15 minutes** (even the correct password is refused while locked).\n\n### How to test\n1. Register a customer first (`POST /api/v1/users`).\n2. Log in with the right password. Expect **200** and tokens. Copy `accessToken` into the Authorize box.\n3. Log in with a wrong password. Expect **401 INVALID_CREDENTIALS**. Try an unknown email: the response is identical.\n4. Fail 5 times in a row, then use the correct password. Still **401** (locked for 15 minutes; a password reset unlocks it).\n5. Hit it 11 times in a minute. Expect **429**.\n6. To try the two-step response, enable it under \"4. Two-step verification\" and log in again.\n",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tokens, or an MFA challenge when two-step verification is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TokenResponse"
                    },
                    {
                      "$ref": "#/components/schemas/MfaChallengeResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS (wrong password, unknown email, or locked account).",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMIT_EXCEEDED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "MFA_UNAVAILABLE: two-step is on but the code email could not be sent. No tokens issued.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/mfa/disable": {
      "post": {
        "tags": [
          "4. Two-step verification"
        ],
        "summary": "Turn off two-step verification",
        "description": "Turns two-step verification off. **Requires a bearer token and the current password.**\n\n### How it works\nThe password is required so that a stolen access token alone cannot weaken the account. Wrong passwords here count towards the account lockout.\n\n### How to test\n1. With two-step on, send a wrong password. Expect **401 INVALID_CREDENTIALS** (still enabled).\n2. Send the correct password. Expect **204**.\n3. Log in: you get tokens directly again.\n",
        "operationId": "disableMfa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Two-step verification is now off."
          },
          "400": {
            "description": "VALIDATION_ERROR.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED or INVALID_CREDENTIALS (wrong password).",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/enable/confirm": {
      "post": {
        "tags": [
          "4. Two-step verification"
        ],
        "summary": "Confirm the code and turn on two-step verification",
        "description": "Turns two-step verification on once the emailed code is proven. **Requires a bearer token.**\n\n### How it works\nThe code must belong to **your** account and to this purpose (a login code cannot be used here, and vice versa). Same rules as login codes: 5 minutes, single use, 5 guesses.\nFrom then on, `POST /api/v1/auth/login` returns `mfaRequired` instead of tokens.\n\n### How to test\n1. After `enable/request`, send the `challengeId` and the emailed `code`. Expect **204**.\n2. Log in again: the response is now `{\"mfaRequired\": true, ...}`.\n3. Send a wrong code: **401 INVALID_MFA_CODE**.\n4. Send a code that belongs to someone else's challenge: also **401**.\n",
        "operationId": "confirmEnableMfa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Two-step verification is now on."
          },
          "400": {
            "description": "VALIDATION_ERROR.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED or INVALID_MFA_CODE.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/enable/request": {
      "post": {
        "tags": [
          "4. Two-step verification"
        ],
        "summary": "Start turning on two-step verification",
        "description": "Emails a confirmation code. **Requires a bearer token.** Limited to 10 requests per minute per IP address (shared with the other MFA calls).\n\n### How it works\nA code is emailed first, so you cannot lock yourself out with an address that does not work. Nothing changes until you confirm the code.\nRequesting again replaces the previous code. If two-step is already on you get `409`.\n\n### How to test\n1. Log in and authorize with the `accessToken`.\n2. Call this endpoint. Expect **200** with a `challengeId`. Copy it.\n3. Open the inbox (Mailpit at http://localhost:8025) and read the code.\n4. Continue with `POST /api/v1/auth/mfa/enable/confirm`.\n5. Without a token: **401**. After enabling, calling this again: **409 MFA_ALREADY_ENABLED**.\n",
        "operationId": "requestEnableMfa",
        "responses": {
          "200": {
            "description": "A code was emailed.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MfaChallengeResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "MFA_ALREADY_ENABLED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "MFA_UNAVAILABLE: the code email could not be sent.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DisableMfaRequest": {
        "type": "object",
        "description": "Confirms the account owner before two-step verification is turned off.",
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "description": "The current account password.",
            "example": "Sup3rSecret",
            "maxLength": 128,
            "minLength": 0
          }
        },
        "required": [
          "password"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error body used by every endpoint.",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-30T16:13:30.942Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Stable machine-readable code to branch on.",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation.",
            "example": "email must be a well-formed email address"
          },
          "path": {
            "type": "string",
            "description": "The request path.",
            "example": "/api/v1/users"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "description": "Credentials used to sign in.",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email used at registration (case-insensitive).",
            "example": "ada.obi@example.com",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "The account password.",
            "example": "Sup3rSecret",
            "maxLength": 128,
            "minLength": 0
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "MfaChallengeResponse": {
        "type": "object",
        "description": "A one-time code was emailed. Submit it with this challengeId to continue.",
        "properties": {
          "mfaRequired": {
            "type": "boolean",
            "description": "Always true: the client must complete a second step.",
            "example": true
          },
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "description": "Identifies the emailed code. Send it with the code.",
            "example": "9636018a-2f90-46a7-add3-2549b9841b5f"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds until the code expires.",
            "example": 300
          }
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "description": "A signed-in session: a short-lived access token and a single-use refresh token.",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT (RS256). Send as 'Authorization: Bearer <token>'. Valid 15 minutes. The 'sub' claim is the customer id.",
            "example": "eyJraWQiOiJmbG93cGF5LTEiLCJhbGciOiJSUzI1NiJ9..."
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Access token lifetime in seconds.",
            "example": 900
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque token valid 7 days. Exchange it at /refresh; each one works once.",
            "example": "SsCed_IPmKTsFYekt7bVScwtaMIB1whQezv9bXeKxrI"
          }
        }
      },
...
      "VerifyOtpRequest": {
        "type": "object",
        "description": "The emailed one-time code together with the challenge it belongs to.",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "description": "The challengeId returned when the code was requested.",
            "example": "9636018a-2f90-46a7-add3-2549b9841b5f"
          },
          "code": {
            "type": "string",
            "description": "The 6-digit code from the email. Valid 5 minutes, single use, 5 attempts.",
            "example": "904826",
            "minLength": 1,
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/v1/auth/login (or /login/verify)."
      }
    }
  }
}